captcha-auto-fill/server/requirements.txt
ddddocr
flask
pillow
